Initial attempt at sizers API, drop iewin
Infernio: Squashed commit and edited so that it doesn't break dev,
consists of the following:

Proof of concept to replace sizers

Inspired by Qt but designed with the intention to fit well with Bash's
current usage of sizers. Every current use of sizers would obviously
need to be rewritten but the goal would be for the resulting code to be
much smaller and/or easier to understand.

Also very very Very much still just a proof of concept. I've only tried
this in heavily refactored code in frames.py.

WIP Replace iewin with webview and rewrite frames.py

Emphasis on WIP! Still a lot of room for improvements.

Finally get rid of the comtypes dependency by dumping iewin and using
wx's webview instead. Also make HtmlCtrl have a fallback non-html mode
in case the user don't have the Webkit dependencies necessary for
the webview. (Right now that is handled mostly outside the HtmlCtrl,
messy)

frames.py also gets a facelift in the process since it is the only place
right now that uses HtmlCtrl. Improvements include:
    * DocBrowser not being bugged (couldn't create new readmes before)
    * less unnecessarily confusing interface in ModChecker (eg. buttons that
    depended on other buttons' state yet not showing this to the user)
    * hopefully less incomprehensible code
    * getting rid of most of the wx imports, altho I'm going for all of
    them

WIP Wrap and move event binding to balt

A proof of concept that isn't particularly pretty but works well enough.
On the other hand, if all wx objects will one day be entirely abstracted
away, this might need to be made entirely different, or made as a part
of some base GUI object.

Either way, making events and event binding not depend on wx will be
tricky.

Bin the SetUAC function in patcher_dialog

As far as I can tell this was only used in one place, and the argument
was always the same thing rendering the whole function entirely
pointless. Also, dumping this function gets rid of a pesky wx
dependency in patcher_dialog too.

Backport my finished HtmlCtrl

From the inf-wx-begone branch, this is the functional WebView-based
HtmlCtrl, applied to the doc browser, mod checker and WryeLog - in other
words, to all places where WB displays HTML.

Note that this assumes we will migrate to wx 2.9+, since this
implementation won't work on 2.8 at all.

Under # 190

Co-authored-by: Infernio <infernio@icloud.com>