Proof of concept to replace sizers
Inspired by Qt but designed with the intention to fit well with Bash's
current usage of sizers. Every current use of sizers would obviously
need to be rewritten but the goal would be for the resulting code to be
much smaller and/or easier to understand.

Also very very Very much still just a proof of concept. I've only tried
this in heavily refactored code in frames.py.

Under # 190