WIP Wrap and move event binding to balt
A proof of concept that isn't particularly pretty but works well enough.
On the other hand, if all wx objects will one day be entirely abstracted
away, this might need to be made entirely different, or made as a part
of some base GUI object.

Either way, making events and event binding not depend on wx will be
tricky.

Under # 190