Replace wx sizers with custom layouts
This removes all traces of wx's sizers except in bash.py.
They are replaced with layouts, a group of classes that
should be abstracted enough to work with not only wx but also other gui
libraries if needed.

The design of the layout API is a mix between wx, qt, and the existing
thin wrappers around the sizers in bash. They are not quite as flexible
as wx's sizers yet, but they should be easier to understand and less
error prone. It's still a WIP however so suggestions for improvements
are very welcome.

Visually, most windows and widgets should not be very different as I
tried to stay with the original design, but in some places there might
be some differences (especially when it comes to spacing and borders).

Utumno - conflict due to 9e222e0ff0fda3f0c0d5b106d7f9b028a1f7791a

Mopy/bash/basher/dialogs.py:

@@@ -294,36 -292,17 +292,54 @@@ class ImportFaceDialog(balt.Dialog)
          self.picture = balt.Picture(self,350,210,scaling=2)
  <<<<<<< HEAD
 +        #--Layout
 +        fgSizer = wx.FlexGridSizer(5,2,2,4)
 +        fgSizer.AddGrowableCol(1,1)
 +        fgSizer.AddMany([
 +            self.nameCheck,
 +            self.nameText,
 +            self.raceCheck,
 +            self.raceText,
 +            self.genderCheck,
 +            self.genderText,
 +            self.statsCheck,
 +            self.statsText,
 +            self.classCheck,
 +            self.classText,
 +            ])
 +        sizer = hSizer(
 +            (self.listBox,1,wx.EXPAND|wx.TOP,4),
 +            (vSizer(
 +                self.picture, vspace(),
 +                (hSizer(
 +                    (fgSizer,1),
 +                    (vSizer(
 +                        (importButton,0,wx.ALIGN_RIGHT),
 +                        vspace(), CancelButton(self),
 +                        )),
 +                    ),0,wx.EXPAND),
 +                ),0,wx.EXPAND|wx.ALL,4),
 +            )
 +        #--Done
 +        if 'ImportFaceDialog' in balt.sizes:
 +            self.SetSizer(sizer)
 +            self.SetSize(balt.sizes['ImportFaceDialog'])
 +        else:
 +            self.SetSizerAndFit(sizer)
  =======
+         GridLayout(border=4, stretch_cols=[0, 1], stretch_rows=[0], items=[
+             # Row 1
+             ((self.listBox, LayoutOptions(row_span=2, fill=True)),
+              (self.picture, LayoutOptions(col_span=2, fill=True))),
+             # Row 2
+             (None,  # note the row_span in the prev row
+              GridLayout(h_spacing=4, v_spacing=2, stretch_cols=[1], items=[
+                  (self.nameCheck, self.nameText),
+                  (self.raceCheck, self.raceText),
+                  (self.genderCheck, self.genderText),
+                  (self.statsCheck, self.statsText),
+                  (self.classCheck, self.classText)]),
+              (VLayout(spacing=4, items=[importButton, CancelButton(self)]),
+               LayoutOptions(h_align=balt.RIGHT, v_align=balt.BOTTOM)))
+         ]).apply_to(self)
  >>>>>>> 51c07705... Replace wx sizers with custom layouts

Kept nycz's version but may need patching

Under # 190