Initial update_ONAM pass
This adds a method to ModInfo that will add / strip ONAMs for the
plugin depending on the detection from the previous commit.

Right now, it only supports stripping ONAM (which works correctly,
tested with xEdit and some mods).

Like the TODO mentions, we should look out for aers and Elminster's
work on ONAMs - depending on how they solve the problem, Bash might
need an 'always-ONAM' mode for Skyrim as well.