Add nightly deployment
Automates deploying nightly builds to the dropbox shared folder and to
each of Wrye Bash's nexus pages.

Dropbox:
 - Uses dropbox's cli oauth to allow access to the shared folder;
 - The oauth process above can be overridden by providing your own
   access token;
 - Will deploy to the root of the shared folder - if previous builds are
   present it will delete them before uploading the new ones;
 - Test/Branch builds are supported by deploying to a subfolder via the
   '--branch' flag.

Nexus:
 - Requires Google Chrome installed and an active nexus session in
   either Chrome, Firefox or Safari;
 - The appropriate chromedriver for the Chrome version is automatically
   downloaded;
 - The required nexus auth cookies ('member_id', 'pass_hash', and 'sid')
   are automatically retrieved for one of the above browsers;
 - Both chromedriver version and auth cookies can be overridden;
 - Selenium will run Chrome headless while deploying to prevent
   accidental user input;
 - If previous builds are present in the Updates section that match the
   current regex they will be deleted and replaced by the newer files.

A log is kept of all verbose output (sensitive information is not
logged). The path to this log can be changed via '--logfile'.

The authentication credentials are stored in plain text to ease
multiple deployments - this can be disabled via '--no-config'.

For testing purposes, a dry run can be executed via '--dry-run'.

The folder containing distributables to deploy can be changed via
'--dist-folder'.

It is recommended to always use 'deploy.py' to deploy - disabling
deployment to either target can be via '--no-dropbox' or '--no-nexus'.