Update to wxPython 3.0.2.0, drop comtypes
Included edit to build.py here.
Also increased recommended Python version to 2.7.16.

Workaround for locale crash

Ran all PNGs through ImageMagick

Fixes the 'iCCP: known incorrect sRGB profile' warning at startup.
wx3 seems to have upgraded to libpng 1.6+, which is much more stringent
about iCCP profiles:
https://stackoverflow.com/questions/22745076/libpng-warning-iccp-known-incorrect-srgb-profile

Command and output are interesting:

  $ find . -iname *.png -print0 | xargs -0 magick mogrify
  mogrify: iCCP: CRC error `./bash/images/loot16.png' @ warning/png.c/MagickPNGWarningHandler/1748.
  mogrify: iCCP: CRC error `./bash/images/loot24.png' @ warning/png.c/MagickPNGWarningHandler/1748.
  mogrify: iCCP: CRC error `./bash/images/loot32.png' @ warning/png.c/MagickPNGWarningHandler/1748.

Fix for crashing Colors dialog

Same problem as in 360a516b56b804e184984b9bd311f3e2195372c8.
Need to check if this is wx3-only or if it's needed in Phoenix as well.
Also need to think about wrapping ComboBox in the wx-begone branch to
fix this in a non-hacky way, no GUI component should ever hard-crash the
application if it's not used correctly :P

Hacky workaround for setCancel crash

Destroys the dialog and creates a similar-looking one with the correct
PD_CAN_ABORT flag set. This is probably the solution we'll have to end
up going with if we want to keep the native ProgressDialog (which we
*really* do, some people have reported *massive* speedups because of it).

Failing to unpickle settings due to _gdi not being resolved as a module

See https://stackoverflow.com/a/13398680/281545

Co-authored-by: MrD <the.ubik@gmail.com>