Fix OBSEIniFile tweaks not displaying correctly
Fixup for 23844a64a2b20aa7685dfb20754dc206ad59d63a, I think. The rest of
that method depended on 'section' being reassigned by the for loop, but
the refactoring changed it to assign to 'section_key', which is then
unused in the rest of the method.

Noticed this while attempting to recreate the 'All Natural.ini'
screenshot in our readme with nightly code. This didn't break applying
of tweaks, thankfully, only the green/yellow line display and scrolling
on double-click.