Update to wxPython 3.0.2.0, drop comtypes RRR
Included edit to build.py here.
Also increased recommended Python version to 2.7.16.

Workaround for locale crash

Fix for crashing Colors dialog

Same problem as in 360a516b56b804e184984b9bd311f3e2195372c8.
Need to check if this is wx3-only or if it's needed in Phoenix as well.
Also need to think about wrapping ComboBox in the wx-begone branch to
fix this in a non-hacky way, no GUI component should ever hard-crash the
application if it's not used correctly :P

Hacky workaround for setCancel crash

Destroys the dialog and creates a similar-looking one with the correct
PD_CAN_ABORT flag set. This is probably the solution we'll have to end
up going with if we want to keep the native ProgressDialog (which we
*really* do, some people have reported *massive* speedups because of it).

Failing to unpickle settings due to _gdi not being resolved as a module

See https://stackoverflow.com/a/13398680/281545

Ellipsize overly long messages in progress dialogs

Reintroduces a feature from wx2.8's progress dialogs, but in a fairly
ugly way (by hardcoding a limit on the length of the displayed string -
this won't scale at all to larger displays).

Small stuff:

Mopy/bash/basher/settings_links.py: shadowing, Settings_Language button
should be disabled on current language (chops off one use of ugly
_bassLang) and rename encodings to more private _plugin_encodings.
Moved the init of bolt.pluginEncoding out of UI code (duh).
Needs work
Mopy/bash/bash.py: changed the message to reflect this is the default
user locale

Hacky half fix for Cancel while refreshing projects not being respected:

So start on Installers tab > Cancel on loading data> wait for pickle
to load (can't really cancell this > go to mods tab > go back to Installers
tab > no refresh triggered !!

Then this works as expected - interupts the loading then going to another
tab and back retriggers the refresh

Arghhh -> time to rewrite progress handling

Fix mod checker crashing when closed and reopened

Was caused by Link.Frame.modChecker never being unset - there was even a
TODO for it in the code.

Fix focus not being lost on TextCtrls

This is precisely the reason why my new event framework Skip()'s by
default. Not skipping by default is incredibly toxic and makes it really
easy to create incredibly strange behavior.

Enough ranting - the problem was that several text fields weren't losing
focus anymore. Easiest to reproduce with the ones on the Mods tab. Made
editing them almost impossible since wx had to idea what to do with the
cursor when multiple text fields were active at the same time.

Use pngcrush to avoid iCCP warning

This implements it for wizards, followup commit does the same for
FOMODs.
Also added a licenses.txt file to the Mopy/bash/compiled folder to
satisfy any requirements for attribution.

Add pngcrush executable

This is pngcrush 1.8.11 32 bit from
https://sourceforge.net/projects/pmt/files/pngcrush-executables

It's the last version that a windows executable was uploaded for, 1.8.12
and 1.8.13 do not have official executables so we'd have to build our
own.

TODO: Once we upgrade everything to 64 bit (also TODO: check again why
we haven't done so yet), we should swap this for the 64 bit version as
well, I guess.

First implementation for wizards

- Only wizards
- Only for archives (projects should be fixed by the author)
- Runs on all PNGs (seems to be *really* slow)
- Fixes the issue, however

Stuck it into archives.py right now, but that's not the right place IMO.

Skip PNGs that don't have incorrect profiles

Speeds it up *significantly*, making it pretty much instant for working
PNGs.

Avoid cmd prompts popping up in standalone build

Whoops, forgot to pass in startupinfo.

Delay progress dialog when unpacking to project

Under wx3, it would shove itself over any warning messages / rename
dialogs we showed. This wasn't really a *problem*, since the user could
just move the progress dialog out of the way, but it was annoying. This
is a fairly painless fix that just asks the questions first and then
creates the progress dialog.

Closes # 15, closes # 458. <--- RRR

Co-authored-by: MrD <the.ubik@gmail.com>