XXX support for backups/duplicate in bak files
Does it make sense to duplicate? Backup makes sense - if anyone
ever is using this
This is broken and unfinished but shows (possible) parts of the
code to change. Instead of overwriting UILIst._new_name override
SaveList.new_name. This code is hacky and cranky and full of wx
and the copy/move infos APIs need revisiting - but
d9923405b584b3ebe375142f3f3288605da4fb65 makes it at least
possible (we needed cosaves filenames for bak)