Process tags from Data/BashTags RRR
The idea is to allow people to add tags to plugins without having to
edit the plugin itself or the LOOT masterlist. An example usage would be
someone creating a guide, which would normally involve the user having
to manually add Bash Tags to plugins. Now, a guide creator can simply
use this system to add tags to plugins, package their 'Data/BashTags'
folder as an archive, and users can use BAIN to install the files,
making the correct tags available for their plugins.

Includes Readme edits and an edit to BAIN to allow it to install files
into this folder. Also made barb back these up - useful if the user
manually created any files there.

And on a completely random note, corrected one occurence of 'Data/Bashed
Patches' to 'Data/Bash Patches' (in the docs for barb, it's talking
about the CSV folder).

Closes # 463 <--- RRR