Add support for non-standard SSE saves RRR
Skyrim SE has an INI setting called uiCompression, which defaults to 2,
where it means 'compress the save with LZ4'. However, users can freely
change this value to 1 or 0, which mean 'compress the save with zlib'
and 'do not compress the save at all', respectively.

This adds support for those other save types, which were previously
causing Wrye Bash to lock up when loading, requiring it to be killed via
the task manager.

The new zlib method also contains a safety check - why not throw it in,
might catch errors before we try to parse nonsense and get nonsense
tracebacks as a result.

Closes # 466 <--- RRR